added main for checking as im going
